requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
reportlab==4.0.4
Pillow==10.0.1
